backend/requirements.txt
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
mysqlclient==2.2.7
python-dotenv==1.0.1
requests==2.32.3
langchain==0.3.20
langchain-community==0.3.19
ctransformers==0.2.27
huggingface-hub==0.29.2
sentence-transformers==3.4.1
pillow==11.1.0
numpy==2.2.3
scikit-learn==1.6.1
scipy==1.15.2
protobuf==5.29.3
PyJWT==2.9.0
uvicorn==0.34.0
